fix: moved the components moved components to the correct directory and converted the files from js to jsx.
