+++ src/components/Loading.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './styles/Loading.css';
 
 class Loading extends React.Component {
   render() {
