+++ src/components/styles/Loading.css
@@ -1,0 +1,8 @@
+h1 {
+  position: absolute;
+  width: 322px;
+  height: 50px;
+  left: 479.45px;
+  top: 403px;
+  font-family:;
+}
