+++ src/pages/Album.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Header from './Header';
+import Header from '../components/Header';
 import getMusics from '../services/musicsAPI';
-import MusicCard from './MusicCard';
+import MusicCard from '../components/MusicCard';
 import { getFavoriteSongs } from '../services/favoriteSongsAPI';
-import Loading from './Loading';
+import Loading from '../components/Loading';
 
 class Album extends React.Component {
   state = {
@@ -16,6 +16,7 @@ class Album extends React.Component {
 
   async componentDidMount() {
     const { match: { params: { id } } } = this.props;
+    console.log('EU SOU UM ID', id);
 
     this.setState({ getFavoriteCheck: true });
