+++ src/pages/Favorites.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './Header';
+import Header from '../components/Header';
 
 class Favorites extends React.Component {
   render() {
