+++ src/pages/ProfileEdit.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './Header';
+import Header from '../components/Header';
 
 class ProfileEdit extends React.Component {
   render() {
